fix: make direct cache materialization portable

+++ tools/usd-pointcloud-convert/main.cpp
@@ -341,7 +341,7 @@ bool BuildCacheDescriptor(
     descriptor.source = {
         canonicalPath.generic_string(),
         size,
-        modified.time_since_epoch().count(),
+        static_cast<std::int64_t>(modified.time_since_epoch().count()),
         contentIdentity};
     descriptor.pluginVersion = "usd-pointcloud-plugins-0.2.2";
     descriptor.parserVersion = IsExtension(inputPath, ".las")
